requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
playwright==1.40.0
google-generativeai==0.3.0
requests==2.31.0
aiohttp==3.9.1
pdfplumber==0.10.3
pandas==2.1.3
openpyxl==3.1.2
pillow==10.1.0
pytesseract==0.3.10
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
backoff==2.2.1
beautifulsoup4==4.12.2
matplotlib==3.7.2
numpy==1.24.3